requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
passlib==1.7.1
flask-login==0.4.1
flask-cors>=3.0
psycopg2-binary>=2.7
pytz>=2020
Flask-SocketIO>=4.3.1
eventlet>=0.29.1
gevent>=20.9.0
gevent-websocket>=0.10.1
netifaces>=0.10.9